refactor(tooL): Move ToolCommand, ToolOutput, ... into tool::run module
