+++ iai-callgrind-runner/src/runner/lib_bench.rs
@@ -13,7 +13,7 @@ use super::meta::Metadata;
 use super::summary::{BaselineKind, BaselineName, BenchmarkKind, BenchmarkSummary, SummaryOutput};
 use super::tool::config::ToolConfigs;
 use super::tool::path::{ToolOutputPath, ToolOutputPathKind};
-use super::tool::RunOptions;
+use super::tool::run::RunOptions;
 use crate::api::{
     EntryPoint, LibraryBenchmarkBench, LibraryBenchmarkConfig, LibraryBenchmarkGroups, RawArgs,
     ValgrindTool,
